Fix the RPM build, a format warning and a coloured fringe on card edges
The %files entries for verify-routing.sh and input-audit.sh had been
inserted into %install as well, because the string I replaced on appears
in both sections. rpm then tried to execute them as shell and the build
failed at %install.

theme.cpp printed a qsizetype through %d.

Effect regions are now inset a pixel from the painted surface. The card's
rounded edge is anti-aliased, so its outermost pixel is only partly
opaque, while the effect region is a hard-edged QRegion covering it
fully. With saturation at 180% the backdrop showing through that pixel is
visibly wrong — over a blue banner it fringes cyan.

AFTER-LOGIN.md was missing an SPDX header, which failed the reuse check.

+++ src/theme.cpp
@@ -141,8 +141,9 @@ void Theme::reload()
         m_name = QFileInfo(path).completeBaseName();
         m_loaded = true;
         watchFile(path);
-        qInfo("glassosd: theme '%s' loaded from %s (%d keys)",
-              qUtf8Printable(m_name), qUtf8Printable(path), m_values.size());
+        qInfo("glassosd: theme '%s' loaded from %s (%lld keys)",
+              qUtf8Printable(m_name), qUtf8Printable(path),
+              static_cast<long long>(m_values.size()));
         break;
     }
